- Github #472 review fixes
