+++ bootstrap/src/main/kotlin/jp/co/soramitsu/bootstrap/dto/IrohaModel.kt
@@ -1,0 +1,23 @@
+package jp.co.soramitsu.bootstrap.dto
+
+import java.security.KeyPair
+import javax.xml.bind.DatatypeConverter
+
+data class IrohaAccountDto(
+    val name: String = "",
+    val domainId: String = "",
+    val creds: List<BlockchainCreds> = listOf()
+)
+
+data class IrohaAccount(val title: String, val domain: String, val keys: HashSet<KeyPair>) :
+    DtoFactory<IrohaAccountDto> {
+    override fun getDTO(): IrohaAccountDto {
+        val credsList = keys.map {
+            BlockchainCreds(
+                DatatypeConverter.printHexBinary(it.private.encoded),
+                DatatypeConverter.printHexBinary(it.public.encoded)
+            )
+        }
+        return IrohaAccountDto(this.title, domain, credsList)
+    }
+}
